pyproject.toml
[build-system]
requires = ["poetry_core>=2.0.0"]
build-backend = "poetry.core.masonry.api"


[project]
name = "DeceptionGame"
version = "0.1.0"
authors = [
  { name = "Matthew Lyle Olson", email = "matthew.lyle.olson@intel.com" },
  { name = "Neale Ratzlaff", email = "neale.ratzlaff@intel.com" },
  { name = "Musashi Hinck", email = "musashi.hinck@intel.com" },
]
description = "A game of deception."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
  "Development Status :: 1 - Planning",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering",
  "Typing :: Typed",
]
dependencies = [
    "fastapi>=0.115.12",
    "fire>=0.7.0",
    "numpy>=2.0.2",
    "openai>=1.74.0",
    "pandas>=2.3.1",
    "pick>=2.4.0",
    "pyyaml>=6.0.2",
    "torch>=2.6.0",
    "transformers>=4.50.2",
    "uvicorn>=0.34.0",
    "anthropic (>=0.54.0,<0.55.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "huggingface-hub (>=0.33.0,<0.34.0)",
    "google-genai (>=1.20.0,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "boto3 (>=1.39.9,<2.0.0)",
    "trueskill (>=0.4.5,<0.5.0)",
]

[project.optional-dependencies]
test = [
  "pytest >=6",
  "pytest-cov >=3",
]
dev = [
  "pytest >=6",
  "pytest-cov >=3",
]
docs = [
  "sphinx>=7.0",
  "myst_parser>=0.13",
  "sphinx_copybutton",
  "sphinx_autodoc_typehints",
  "furo>=2023.08.17",
]


[project.urls]
Homepage = "https://github.com/intel-sandbox/iaia.DeceptionGame"
"Bug Tracker" = "https://github.com/intel-sandbox/iaia.DeceptionGame/issues"
Discussions = "https://github.com/intel-sandbox/iaia.DeceptionGame/discussions"
Changelog = "https://github.com/intel-sandbox/iaia.DeceptionGame/releases"

[tool.poetry.group.test.dependencies]
pytest-cov = ">= 3"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[tool.uv]
dev-dependencies = [
  "black>=25.1.0",
  "deceptiongame[test]",
  "pytest>=8.3.5",
]


[tool.pytest.ini_options]
minversion = "6.0"
addopts = ["-ra", "--showlocals", "--strict-markers", "--strict-config"]
xfail_strict = true
filterwarnings = [
  "error",
]
log_cli_level = "INFO"
testpaths = [
  "tests",
]


[tool.coverage]
run.source = ["deceptiongame"]
report.exclude_also = [
  '\.\.\.',
  'if typing.TYPE_CHECKING:',
]

[tool.mypy]
files = ["src", "tests"]
python_version = "3.9"
warn_unused_configs = true
strict = true
enable_error_code = ["ignore-without-code", "redundant-expr", "truthy-bool"]
warn_unreachable = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = "deceptiongame.*"
disallow_untyped_defs = true
disallow_incomplete_defs = true


[tool.ruff]
target-version = "py39"

[tool.ruff.lint]
extend-select = [
  "ARG",      # flake8-unused-arguments
  "B",        # flake8-bugbear
  "C4",       # flake8-comprehensions
  "EM",       # flake8-errmsg
  "EXE",      # flake8-executable
  "G",        # flake8-logging-format
  "I",        # isort
  "ICN",      # flake8-import-conventions
  "NPY",      # NumPy specific rules
  "PD",       # pandas-vet
  "PGH",      # pygrep-hooks
  "PIE",      # flake8-pie
  "PL",       # pylint
  "PT",       # flake8-pytest-style
  "PTH",      # flake8-use-pathlib
  "RET",      # flake8-return
  "RUF",      # Ruff-specific
  "SIM",      # flake8-simplify
  "T20",      # flake8-print
  "UP",       # pyupgrade
  "YTT",      # flake8-2020
]
ignore = [
  "PLR09",    # Too many <...>
  "PLR2004",  # Magic value used in comparison
]
isort.required-imports = ["from __future__ import annotations"]
# Uncomment if using a _compat.typing backport
# typing-modules = ["deceptiongame._compat.typing"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["T20"]
"noxfile.py" = ["T20"]


[tool.pylint]
py-version = "3.9"
ignore-paths = [".*/_version.py"]
reports.output-format = "colorized"
similarities.ignore-imports = "yes"
messages_control.disable = [
  "design",
  "fixme",
  "line-too-long",
  "missing-module-docstring",
  "missing-function-docstring",
  "wrong-import-position",
]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310"]
